Move depend_on_either_json_or_form_data function from tool_shed to galaxy
