+++ lib/galaxy/webapps/galaxy/api/__init__.py
@@ -4,6 +4,7 @@ This module *does not* contain API routes. It exclusively contains dependencies 
 
 import inspect
 from enum import Enum
+from json import JSONDecodeError
 from string import Template
 from typing import (
     Any,
@@ -26,6 +27,7 @@ from fastapi import (
     APIRouter,
     Form,
     Header,
+    HTTPException,
     Query,
     Request,
     Response,
@@ -108,6 +110,8 @@ DependsOnApp = cast(StructuredApp, Depends(get_app_with_request_session))
 
 
 T = TypeVar("T")
+
+B = TypeVar("B", bound=BaseModel)
 
 
 class GalaxyTypeDepends(Depends):
@@ -602,3 +606,26 @@ def search_query_param(model_name: str, tags: list, free_text_fields: list) -> O
         title="Search query.",
         description=description,
     )
+
+
+# TODO: mypy complains that Depends(get_body) is returned B is expected.
+# def depend_on_either_json_or_form_data(model: Type[B]) -> B:
+def depend_on_either_json_or_form_data(model: Type[B]):
+    async def get_body(request: Request):
+        content_type = request.headers.get("Content-Type")
+        if content_type is None:
+            raise HTTPException(status_code=400, detail="No Content-Type provided!")
+        elif content_type == "application/json":
+            try:
+                return model(**await request.json())
+            except JSONDecodeError:
+                raise HTTPException(status_code=400, detail="Invalid JSON data")
+        elif content_type == "application/x-www-form-urlencoded" or content_type.startswith("multipart/form-data"):
+            try:
+                return model(**await request.form())
+            except Exception:
+                raise HTTPException(status_code=400, detail="Invalid Form data")
+        else:
+            raise HTTPException(status_code=400, detail="Content-Type not supported!")
+
+    return Depends(get_body)
